Created new component to transition from using buttons to toggle generation, to tabs, and added generations 6 and 7.
