+++ Results.js
@@ -51,116 +51,134 @@ const Results = props => {
   const flatlistRef = useRef();
 
   useEffect(() => {
-    axios.get(url).then(res => {
-      if (props.generation === 3) {
-        // gen 2
-        let pokeList = [];
-        for (let i = 0; i < res.data.pokemon_entries.length; i++) {
-          if (res.data.pokemon_entries[i].entry_number > 151) {
-            pokeList.push(res.data.pokemon_entries[i]);
+    if (props.generation === 12) {
+      // gen 6 odd edge case with pokedex separation in pokeapi
+      let pokeList = [];
+      for (let i = 12; i < 15; i++) {
+        axios.get("https://pokeapi.co/api/v2/pokedex/" + i + "/").then(res => {
+          for (let i = 0; i < res.data.pokemon_entries.length; i++) {
+            const entryId = parseInt(
+              res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
+            );
+            if (entryId > 649) {
+              pokeList.push({
+                ...res.data.pokemon_entries[i],
+                entryId: entryId
+              });
+            }
           }
-        }
-        setPokemon(pokeList);
-      } else {
-        // gen 1
-        setPokemon(res.data.pokemon_entries);
+        });
       }
-    });
-  }, []);
-
-  useEffect(() => {
-    axios.get(url).then(res => {
-      if (props.generation === 3) {
-        // gen 2
-        let pokeList = [];
-        for (let i = 0; i < res.data.pokemon_entries.length; i++) {
-          const entryId = parseInt(
-            res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
+      setPokemon(pokeList.sort((a, b) => (a.entryId > b.entryId ? 1 : -1)));
+    } else {
+      axios.get(url).then(res => {
+        if (props.generation === 3) {
+          // gen 2
+          let pokeList = [];
+          for (let i = 0; i < res.data.pokemon_entries.length; i++) {
+            const entryId = parseInt(
+              res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
+            );
+            if (entryId > 151) {
+              pokeList.push({
+                ...res.data.pokemon_entries[i],
+                entryId: entryId
+              });
+            }
+          }
+          setPokemon(pokeList.sort((a, b) => (a.entryId > b.entryId ? 1 : -1)));
+        } else if (props.generation === 4) {
+          // gen 3
+          let pokeList = [];
+          for (let i = 0; i < res.data.pokemon_entries.length; i++) {
+            const entryId = parseInt(
+              res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
+            );
+            if (entryId > 251) {
+              pokeList.push({
+                ...res.data.pokemon_entries[i],
+                entryId: entryId
+              });
+            }
+          }
+          setPokemon(pokeList.sort((a, b) => (a.entryId > b.entryId ? 1 : -1)));
+        } else if (props.generation === 5) {
+          // gen 4
+          let pokeList = [];
+          for (let i = 0; i < res.data.pokemon_entries.length; i++) {
+            const entryId = parseInt(
+              res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
+            );
+            if (entryId > 386) {
+              pokeList.push({
+                ...res.data.pokemon_entries[i],
+                entryId: entryId
+              });
+            }
+          }
+          const darkrai = {
+            entryId: 491,
+            pokemon_species: {
+              name: "Darkrai",
+              url: "https://pokeapi.co/api/v2/pokemon-species/491/"
+            }
+          };
+          const shaymin = {
+            entryId: 492,
+            pokemon_species: {
+              name: "Shaymin",
+              url: "https://pokeapi.co/api/v2/pokemon-species/492/"
+            }
+          };
+          const arceus = {
+            entryId: 493,
+            pokemon_species: {
+              name: "Arceus",
+              url: "https://pokeapi.co/api/v2/pokemon-species/493/"
+            }
+          };
+          setPokemon(
+            [...pokeList, darkrai, shaymin, arceus].sort((a, b) =>
+              a.entryId > b.entryId ? 1 : -1
+            )
           );
-          if (entryId > 151) {
-            pokeList.push({
-              ...res.data.pokemon_entries[i],
-              entryId: entryId
-            });
+        } else if (props.generation === 8) {
+          // gen 5
+          let pokeList = [];
+          for (let i = 0; i < res.data.pokemon_entries.length; i++) {
+            const entryId = parseInt(
+              res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
+            );
+            if (entryId > 494) {
+              pokeList.push({
+                ...res.data.pokemon_entries[i],
+                entryId: entryId
+              });
+            }
           }
+          setPokemon(pokeList.sort((a, b) => (a.entryId > b.entryId ? 1 : -1)));
+        } else if (props.generation === 16) {
+          // gen 7
+          let pokeList = [];
+          for (let i = 0; i < res.data.pokemon_entries.length; i++) {
+            const entryId = parseInt(
+              res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
+            );
+            if (entryId > 721) {
+              pokeList.push({
+                ...res.data.pokemon_entries[i],
+                entryId: entryId
+              });
+            }
+          }
+          setPokemon(pokeList.sort((a, b) => (a.entryId > b.entryId ? 1 : -1)));
+        } else {
+          // gen 1
+          setPokemon(res.data.pokemon_entries);
         }
-        setPokemon(pokeList.sort((a, b) => (a.entryId > b.entryId ? 1 : -1)));
-      } else if (props.generation === 4) {
-        // gen 3
-        let pokeList = [];
-        for (let i = 0; i < res.data.pokemon_entries.length; i++) {
-          const entryId = parseInt(
-            res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
-          );
-          if (entryId > 251) {
-            pokeList.push({
-              ...res.data.pokemon_entries[i],
-              entryId: entryId
-            });
-          }
-        }
-        setPokemon(pokeList.sort((a, b) => (a.entryId > b.entryId ? 1 : -1)));
-      } else if (props.generation === 5) {
-        // gen 4
-        let pokeList = [];
-        for (let i = 0; i < res.data.pokemon_entries.length; i++) {
-          const entryId = parseInt(
-            res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
-          );
-          if (entryId > 386) {
-            pokeList.push({
-              ...res.data.pokemon_entries[i],
-              entryId: entryId
-            });
-          }
-        }
-        const darkrai = {
-          entryId: 491,
-          pokemon_species: {
-            name: "Darkrai",
-            url: "https://pokeapi.co/api/v2/pokemon-species/491/"
-          }
-        };
-        const shaymin = {
-          entryId: 492,
-          pokemon_species: {
-            name: "Shaymin",
-            url: "https://pokeapi.co/api/v2/pokemon-species/492/"
-          }
-        };
-        const arceus = {
-          entryId: 493,
-          pokemon_species: {
-            name: "Arceus",
-            url: "https://pokeapi.co/api/v2/pokemon-species/493/"
-          }
-        };
-        setPokemon(
-          [...pokeList, darkrai, shaymin, arceus].sort((a, b) =>
-            a.entryId > b.entryId ? 1 : -1
-          )
-        );
-      } else if (props.generation === 8) {
-        // gen 5
-        let pokeList = [];
-        for (let i = 0; i < res.data.pokemon_entries.length; i++) {
-          const entryId = parseInt(
-            res.data.pokemon_entries[i].pokemon_species.url.slice(42, -1)
-          );
-          if (entryId > 494) {
-            pokeList.push({
-              ...res.data.pokemon_entries[i],
-              entryId: entryId
-            });
-          }
-        }
-        setPokemon(pokeList.sort((a, b) => (a.entryId > b.entryId ? 1 : -1)));
-      } else {
-        // gen 1
-        setPokemon(res.data.pokemon_entries);
-      }
-    });
-    flatlistRef.current.scrollToOffset({ animated: true, offset: 0 });
+      });
+    }
+    // flatlistRef.current.scrollToOffset({ animated: true, offset: 0 });
   }, [props.generation]);
 
   return (
